Solving puzzle 2 for day 4!
Fixing bug with checkForBingo logic where if we had a vertical bingo, it would get marked and then overridden because the for loop would continue over and remark hasVertical as false. Adding an if statement to the for loop so if we have a vertical bingo, we stop looking for other vertical bingo matches.